Split up controller
Resolves #112
